+++ plugins/__init__.py
@@ -15,7 +15,6 @@
 
     Contact: code@inmanta.com
 """
-
 import base64
 import hashlib
 import importlib
@@ -1079,8 +1078,32 @@ def validate_type(
         * Define a vlan_id type which represent a valid vlan ID (0-4,095):
 
           typedef vlan_id as number matching std::validate_type("pydantic.conint", self, {"ge": 0, "le": 4095})
+    """
+    try:
+        import inmanta.validation_type
+    except ModuleNotFoundError:
+        # We are running against a version of inmanta-core that doesn't have the validation_type method yet.
+        # Fallback to the old implementation.
+        return _validate_type_legacy(fq_type_name, value, validation_parameters)
+    else:
+        # Use validate_type implementation from inmanta-core
+        unwrapped_value = DynamicProxy.unwrap(value)
+        if isinstance(unwrapped_value, NoneValue):
+            unwrapped_value = None
+        try:
+            inmanta.validation_type.validate_type(
+                fq_type_name, unwrapped_value, validation_parameters
+            )
+        except (pydantic.ValidationError, ValueError):
+            return False
+        return True
 
 
+def _validate_type_legacy(
+    fq_type_name: "string", value: "any", validation_parameters: "dict" = None
+) -> "bool":
+    """
+    This method contains the old implementation of the validate_type plugin for backwards compatibility reason.
     """
     if not (
         fq_type_name.startswith("pydantic.")